+++ NEWS.md
@@ -4,6 +4,8 @@
 - `tensor_image_display` and `tensor_image_browse` now accept all tensor_image dtypes. (#115, @cregouby) 
 - fix `transform_affine` help to remove confusion with `transforme_random_affine` help (#116, @cregouby)
 - add message translation in french (#112, @cregouby)
+- Added support for the Fashion-MNIST dataset. (#148, @koshtiakanksha)
+- Added support for the MS COCO Captions dataset. (#141, @koshtiakanksha)
 - Added the Fashion-MNIST dataset. (#148, @koshtiakanksha)
 - Added 3 EuroSAT datasets (#126 @cregouby)
 - Added `qmnist_dataset()` – a dataset loader for the QMNIST dataset (#153, @DerrickUnleashed)